pyproject.toml
[project]
name = "paimon-quickstart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "duckdb>=1.5.1",
    "httpx>=0.28.1",
    "kafka-python>=2.3.0",
    "pypaimon>=1.3.1",
    "sseclient-py>=1.9.0",
]
